chore: rename remaining types
