+++ stoix/configs/arch/anakin.yaml
@@ -2,7 +2,7 @@
 
 # --- Training ---
 seed: 42  # RNG seed.
-num_envs: 128  # Number of vectorised environments per device.
+num_envs: 2  # Number of vectorised environments per device.
 total_timesteps: 1e7 # Set the total environment steps.
 # If unspecified, it's derived from num_updates; otherwise, num_updates adjusts based on this value.
 num_updates: ~ # Number of updates
