+++ stoix/systems/ppo/ff_ppo.py
@@ -18,7 +18,7 @@ from stoix.base_types import ActorApply, CriticApply, ExperimentOutput, LearnerF
 from stoix.evaluator import evaluator_setup, get_distribution_act_fn
 from stoix.networks.base import FeedForwardActor as Actor
 from stoix.networks.base import FeedForwardCritic as Critic
-from stoix.systems.ppo.types import (
+from stoix.systems.ppo.ppo_types import (
     ActorCriticOptStates,
     ActorCriticParams,
     LearnerState,
