+++ stoix/systems/ppo/rec_ppo.py
@@ -17,7 +17,7 @@ from rich.pretty import pprint
 from stoix.base_types import ExperimentOutput, LearnerFn, RecActorApply, RecCriticApply
 from stoix.evaluator import evaluator_setup, get_rec_distribution_act_fn
 from stoix.networks.base import RecurrentActor, RecurrentCritic, ScannedRNN
-from stoix.systems.ppo.types import (
+from stoix.systems.ppo.ppo_types import (
     ActorCriticOptStates,
     ActorCriticParams,
     HiddenStates,
