+++ stoix/systems/q_learning/ff_dqn_reg.py
@@ -20,7 +20,7 @@ from rich.pretty import pprint
 from stoix.base_types import ActorApply, ExperimentOutput, LearnerFn, LogEnvState
 from stoix.evaluator import evaluator_setup, get_distribution_act_fn
 from stoix.networks.base import FeedForwardActor as Actor
-from stoix.systems.q_learning.types import DQNLearnerState, QsAndTarget, Transition
+from stoix.systems.q_learning.dqn_types import DQNLearnerState, QsAndTarget, Transition
 from stoix.utils import make_env as environments
 from stoix.utils.checkpointing import Checkpointer
 from stoix.utils.jax_utils import unreplicate_batch_dim, unreplicate_n_dims
