+++ stoix/systems/sac/ff_sac.py
@@ -22,7 +22,7 @@ from stoix.evaluator import evaluator_setup, get_distribution_act_fn
 from stoix.networks.base import CompositeNetwork
 from stoix.networks.base import FeedForwardActor as Actor
 from stoix.networks.base import MultiNetwork
-from stoix.systems.q_learning.types import QsAndTarget, Transition
+from stoix.systems.q_learning.dqn_types import QsAndTarget, Transition
 from stoix.systems.sac.sac_types import (
     ContinuousQApply,
     SACLearnerState,
